Make autocomplete a component
